requirements.txt
asn1crypto
attrs
Babel
beautifulsoup4
cached-property
cbor2
certifi
cffi
chardet
cryptography
Cython
decorator
docopt
docutils
et_xmlfile
exceptiongroup
freezegun
geoip2
gevent
greenlet
idna
iniconfig
ipaddress
isodate
Jinja2
libsass
lxml
MarkupSafe
maxminddb
monero
num2words
numpy
ofxparse
opencv-python
openpyxl
packaging
passlib
Pillow
pip
platformdirs
pluggy
polib
psutil
psycopg2
pyasn1
pyasn1_modules
pycparser
pycryptodomex
Pygments
PyMuPDF
PyNaCl
pyOpenSSL
PyPDF2
pyserial
PySocks
pytesseract
pytest
pytest-odoo
python-dateutil
python-ldap
python-stdnum
pytz
pyusb
qrcode
reportlab
requests
requests-file
requests-toolbelt
rjsmin
setuptools
six
soupsieve
sphinx 
sphinx-autobuild 
sphinx_rtd_theme
tomli
typing_extensions
urllib3
varint
vobject
Werkzeug
wheel
xlrd
XlsxWriter
xlwt
zeep
zope.event
zope.interface
sphinx 
sphinx-autobuild 
sphinx_rtd_theme

